+++ schloesser/.gitignore
@@ -1,3 +1,3 @@
-appfs/*
-appfs/
+venv/*
+venv/
