+++ schloesser/exercise05/.gitignore
@@ -1,4 +1,4 @@
 *.gph*
 
 *.md
-
+*.ipynb
